various fixes, related to 32-bits.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@331 3eaf19af-ecc0-40f6-b43f-672aa0c71c71

+++ linalg/blockmatrix.c
@@ -54,7 +54,7 @@ void blockmatrix_print(blockmatrix b, const char *vname)
 		    mat64_ptr bl = b->mb[i / 64 + (j / 64) * b->stride];
 		    if (j + jj)
 			fprintf(f,",");
-		    fprintf(f,"%"PRIu64, (bl[ii] >> jj) & 1UL);
+		    fprintf(f,"%"PRIu64, (bl[ii] >> jj) & ((uint64_t)1));
 		}
 	    }
 	}
